+++ packages/form/src/components/TimePicker/index.tsx
@@ -13,17 +13,15 @@ const valueType = 'time';
 const ProFormTimePicker: React.FC<ProFormItemProps<DatePickerProps>> = ({
   fieldProps,
   proFieldProps,
-}) => {
-  return (
-    <ProField
-      text={fieldProps?.value || ''}
-      mode="edit"
-      fieldProps={fieldProps}
-      valueType={valueType}
-      {...proFieldProps}
-    />
-  );
-};
+}) => (
+  <ProField
+    text={fieldProps?.value || ''}
+    mode="edit"
+    fieldProps={fieldProps}
+    valueType={valueType}
+    {...proFieldProps}
+  />
+);
 
 export default createField<ProFormItemProps<DatePickerProps>>(ProFormTimePicker, {
   customLightMode: true,