+++ packages/form/src/demos/steps-from.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import ProForm, {
-  StepsForm,
+  StepsFrom,
   ProFormText,
   ProFormDatePicker,
   ProFormDateTimePicker,
-  ProFormDateRangePicker,
-  ProFormTimePicker,
-  ProFormDateTimeRangePicker,
+  ProFormSelect,
   ProFormTextArea,
   ProFormCheckbox,
 } from '@ant-design/pro-form';
+
+import ProCard from '@ant-design/pro-card';
 
 const waitTime = (time: number = 100) => {
   return new Promise((resolve) => {
@@ -21,14 +21,18 @@ const waitTime = (time: number = 100) => {
 
 export default () => {
   return (
-    <>
-      <StepsForm onFinish={(values) => console.log(values)}>
-        <StepsForm.StepForm
-          name="base"
-          title="基础信息"
-          validateMessages={{
+    <ProCard>
+      <StepsFrom
+        onFinish={(values) => console.log(values)}
+        formProps={{
+          validateMessages: {
             required: '此项为必填项',
-          }}
+          },
+        }}
+      >
+        <StepsFrom.StepFrom
+          name="base"
+          title="创建实验"
           onFinish={async () => {
             await waitTime(2000);
             return true;
@@ -36,37 +40,77 @@ export default () => {
         >
           <ProFormText
             name="name"
-            label="名称"
+            label="实验名称"
             tip="最长为 24 位，用于标定的唯一 id"
             placeholder="请输入名称"
             rules={[{ required: true }]}
           />
-          <ProFormTextArea
-            name="remark"
-            label="备注"
-            placeholder="请输入备注"
-            tip="最长为 24 位，用于标定的唯一 id,最长为 24 位，用于标定的唯一 id,最长为 24 位，用于标定的唯一 id"
-          />
           <ProFormDatePicker name="date" label="日期" />
-
-          <ProForm.Group title="这是一个分组">
-            <ProFormDateTimePicker name="dateTime" label="日期时间" />
-            <ProFormDateRangePicker name="dateRange" label="日期区间" />
+          <ProForm.Group title="时间选择">
+            <ProFormDateTimePicker name="dateTime" label="开始时间" />
+            <ProFormDatePicker name="date" label="结束时间" />
           </ProForm.Group>
-        </StepsForm.StepForm>
-        <StepsForm.StepForm name="checkbox" title="行业分布">
+          <ProFormTextArea name="remark" label="备注" width="l" placeholder="请输入备注" />
+        </StepsFrom.StepFrom>
+        <StepsFrom.StepFrom name="checkbox" title="设置参数">
           <ProFormCheckbox.Group
             name="checkbox"
-            layout="vertical"
-            label="行业分布"
-            options={['农业', '制造业', '互联网']}
+            label="迁移类型"
+            width="l"
+            options={['结构迁移', '全量迁移', '增量迁移', '全量校验']}
           />
-        </StepsForm.StepForm>
-        <StepsForm.StepForm name="time" title="时间选择">
-          <ProFormDateTimeRangePicker name="dateTimeRange" label="日期时间区间" />
-          <ProFormTimePicker name="time" label="时间选择" placeholder="请选择" />
-        </StepsForm.StepForm>
-      </StepsForm>
-    </>
+          <ProForm.Group>
+            <ProFormText name="dbname" label="业务 DB 用户名" />
+            <ProFormDatePicker name="datetime" label="记录保存时间" width="s" />
+            <ProFormCheckbox.Group
+              name="checkbox"
+              label="迁移类型"
+              options={['完整 LOB', '不同步 LOB', '受限制 LOB']}
+            />
+          </ProForm.Group>
+        </StepsFrom.StepFrom>
+        <StepsFrom.StepFrom name="time" title="发布实验">
+          <ProFormCheckbox.Group
+            name="checkbox"
+            label="部署单元"
+            rules={[
+              {
+                required: true,
+              },
+            ]}
+            options={['部署单元1', '部署单元2', '部署单元3']}
+          />
+          <ProFormSelect
+            label="部署分组策略"
+            name="remark"
+            rules={[
+              {
+                required: true,
+              },
+            ]}
+            initialValue="1"
+            options={[
+              {
+                value: '1',
+                label: '策略一',
+              },
+              { value: '2', label: '策略二' },
+            ]}
+          />
+          <ProFormSelect
+            label="Pod 调度策略"
+            name="remark"
+            initialValue="2"
+            options={[
+              {
+                value: '1',
+                label: '策略一',
+              },
+              { value: '2', label: '策略二' },
+            ]}
+          />
+        </StepsFrom.StepFrom>
+      </StepsFrom>
+    </ProCard>
   );
 };