+++ packages/layout/src/SiderMenu/index.tsx
@@ -18,6 +18,7 @@ const SiderMenuWrapper: React.FC<SiderMenuProps> = (props) => {
     style,
     className,
     hide,
+    getContainer,
     prefixCls,
   } = props;
   const { setFlatMenuKeys } = MenuCounter.useContainer();
@@ -59,13 +60,15 @@ const SiderMenuWrapper: React.FC<SiderMenuProps> = (props) => {
         height: '100vh',
         ...style,
       }}
+      getContainer={getContainer}
       width={siderWidth}
-      bodyStyle={{ height: '100vh', padding: 0 }}
+      bodyStyle={{ height: '100vh', padding: 0, display: 'flex', flexDirection: 'row' }}
     >
       <SiderMenu
         {...omitProps}
         className={classNames(`${prefixCls}-sider`, className)}
         collapsed={isMobile ? false : collapsed}
+        splitMenus={false}
       />
     </Drawer>
   ) : (