+++ packages/table/src/Table.tsx
@@ -35,10 +35,11 @@ import {
   ProSchema,
   ProSchemaComponentTypes,
   LabelIconTip,
-  pickUndefinedAndArray,
+  omitUndefinedAndEmptyArr,
   ProCoreActionType,
-  pickUndefined,
   SearchTransformKeyFn,
+  isNil,
+  omitUndefined,
 } from '@ant-design/pro-utils';
 
 import useFetchData, { RequestData } from './useFetchData';
@@ -47,7 +48,6 @@ import Toolbar, { OptionConfig, ToolBarProps } from './component/ToolBar';
 import Alert from './component/Alert';
 import FormSearch, { SearchConfig, TableFormItem } from './Form';
 import {
-  checkUndefinedOrNull,
   genColumnKey,
   genCopyable,
   genEllipsis,
@@ -365,7 +365,7 @@ const columnRender = <T, U = any>({
     }
     return renderDom as React.ReactNode;
   }
-  return checkUndefinedOrNull(dom) ? dom : null;
+  return !isNil(dom) ? dom : null;
 };
 
 /**
@@ -445,7 +445,7 @@ const genColumnList = <T, U = {}>(
         render: (text: any, row: T, index: number) =>
           columnRender<T>({ item, text, row, index, columnEmptyText, counter }),
       };
-      return pickUndefinedAndArray(tempColumns);
+      return omitUndefinedAndEmptyArr(tempColumns);
     })
     .filter((item) => !item.hideInTable) as unknown) as Array<
     ColumnsType<T>[number] & {
@@ -669,8 +669,8 @@ const ProTable = <T extends {}, U extends ParamsType>(
   useDeepCompareEffect(() => {
     if (propsPagination && (propsPagination.current || propsPagination.pageSize)) {
       action.setPageInfo({
-        pageSize: propsPagination.pageSize,
-        page: propsPagination.current,
+        pageSize: propsPagination.pageSize || action.current,
+        page: propsPagination.current || action.pageSize,
       });
     }
   }, [propsPagination && propsPagination.pageSize, propsPagination && propsPagination.current]);
@@ -771,7 +771,7 @@ const ProTable = <T extends {}, U extends ParamsType>(
           rest.onChange(changePagination, filters, sorter, extra);
         }
         // 制造筛选的数据
-        setProFilter(pickUndefinedAndArray<any>(filters));
+        setProFilter(omitUndefinedAndEmptyArr<any>(filters));
 
         // 制造一个排序的数据
         if (Array.isArray(sorter)) {
@@ -783,9 +783,9 @@ const ProTable = <T extends {}, U extends ParamsType>(
               [`${value.field}`]: value.order,
             };
           }, {});
-          setProSort(pickUndefined<any>(data));
+          setProSort(omitUndefined<any>(data));
         } else {
-          setProSort(pickUndefined({ [`${sorter.field}`]: sorter.order as SortOrder }));
+          setProSort(omitUndefined({ [`${sorter.field}`]: sorter.order as SortOrder }));
         }
       }}
     />