+++ tests/setupTests.js
@@ -15,6 +15,8 @@ if (typeof window !== 'undefined') {
   // ref: https://github.com/ant-design/ant-design/issues/18774
   if (!window.matchMedia) {
     Object.defineProperty(global.window, 'matchMedia', {
+      writable: true,
+      configurable: true,
       value: jest.fn(() => ({
         matches: false,
         addListener: jest.fn(),